mtc: Use workaround for FSP bug
The clock change code uses the FSP source clock variable that is normally not passed around the boot stages.
Kernel expects it to be false. Not accounting for that, can lead to system hangs on clock change.
This workaround checks its status and if true does 2 clock switches to return it to false.